Fix obscure encoding bug in player
In windows, when a command is called using subprocess.call() and the
arguments to the program contain unicode characters, the command needs to
be encoded in the locales preferred encoding, otherwise an encoding error
will be raised. This commit fixes this by encoding all the strings in the
command list in the system's preferred encoding.